+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Change Log
 
+- [December, 2023] Voice Interpreter app added
 - [December, 2023] More responsive text-to-speech in auto-speech mode
 - [December, 2023] Language Practice Plus app updated
 - [November, 2023] Image generator app updated to support dall-e-3.